Moved text generation errors
